Update style.scss
did a lot. removed height variable and set one variable that will pass to build grid space. fine tuned the before and after size and spacing for squares. made a mixin to pass to before and after for the sqaure and the background square. added a box shadow to square to make it really pop